refactor: rename routes components
