+++ frontend/svelte/src/routes/Auth.svelte
