+++ frontend/svelte/src/routes/Canisters.svelte
