+++ frontend/svelte/src/routes/Neurons.svelte
