+++ frontend/svelte/src/routes/Voting.svelte
